Decompose ChallengeForm into preset, fields, and actions sub-views
Was 484 lines doing five things at once. Now:
  - components/PresetPicker.jsx (95) -- create-flow Quick Start dropdown
    plus a preview table of the actions the preset will spawn
  - components/ChallengeFieldsSection.jsx (129) -- the field grid
    shared by both create and edit
  - components/ActionsEditor.jsx (210) -- edit-flow actions table with
    its own add/edit dialog (owns the dialog state since nothing else
    uses it)
  - ChallengeForm.jsx (182) -- form state, submit pipeline, layout

Each new file gets a header docblock and JSDoc on its props per the
Cohesive Comment Standard. Behavior unchanged. Refs #42.
