+++ src/admin-app/src/features/challenges/ChallengeForm.jsx
@@ -1,48 +1,38 @@
+/**
+ * ChallengeForm -- create or edit a challenge.
+ *
+ * Renders three sub-views, all under one route:
+ *   - PresetPicker: only on create, lets the user pre-fill from a template
+ *   - ChallengeFieldsSection: the field grid
+ *   - ActionsEditor: only on edit, manages the challenge's actions
+ *
+ * This component owns the form state and the submit pipeline; the
+ * sub-components are pure presentational.
+ */
+
 import { useEffect, useState } from "react";
 
 import { useNavigate, useParams } from "react-router-dom";
 
-import AddIcon from "@mui/icons-material/Add";
 import ArrowBackIcon from "@mui/icons-material/ArrowBack";
-import DeleteIcon from "@mui/icons-material/Delete";
-import EditIcon from "@mui/icons-material/Edit";
-import {
-  Box,
-  Typography,
-  TextField,
-  MenuItem,
-  Button,
-  Stack,
-  Paper,
-  Grid,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
-  IconButton,
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  DialogActions,
-} from "@mui/material";
+import { Box, Button, Paper, Stack, Typography } from "@mui/material";
 
 import {
-  getChallengeById,
-  createChallenge,
-  updateChallenge,
   CATEGORIES,
   CHALLENGE_STATUSES,
-  getActionsByChallenge,
   createAction,
-  updateAction,
-  deleteAction,
+  createChallenge,
+  getActionsByChallenge,
+  getChallengeById,
   getGroups,
   getPresets,
   logActivity,
+  updateChallenge,
 } from "../../data/api";
 import { useAuth } from "../auth/AuthContext";
+import ActionsEditor from "./components/ActionsEditor";
+import ChallengeFieldsSection from "./components/ChallengeFieldsSection";
+import PresetPicker from "./components/PresetPicker";
 
 const EMPTY_FORM = {
   name: "",
@@ -55,27 +45,18 @@ const EMPTY_FORM = {
   groupId: "",
 };
 
-const EMPTY_ACTION = {
-  name: "",
-  description: "",
-  category: CATEGORIES[0],
-  points: 5,
-};
-
 export default function ChallengeForm() {
   const { id } = useParams();
   const isEdit = Boolean(id);
   const navigate = useNavigate();
   const { user } = useAuth();
+
   const [form, setForm] = useState(EMPTY_FORM);
   const [actions, setActions] = useState([]);
   const [groups, setGroups] = useState([]);
   const [presets, setPresets] = useState([]);
   const [selectedPresetId, setSelectedPresetId] = useState("");
   const [presetActions, setPresetActions] = useState([]);
-  const [actionDialogOpen, setActionDialogOpen] = useState(false);
-  const [editingAction, setEditingAction] = useState(null);
-  const [actionForm, setActionForm] = useState(EMPTY_ACTION);
 
   useEffect(() => {
     getGroups().then(setGroups);
@@ -93,19 +74,19 @@ export default function ChallengeForm() {
     const pid = e.target.value;
     setSelectedPresetId(pid);
     const preset = presets.find((p) => p.id === pid);
-    if (preset) {
-      setForm((prev) => ({
-        ...prev,
-        name: preset.name,
-        description: preset.description,
-        category: preset.category,
-        theme: preset.theme,
-        status: preset.status || prev.status,
-      }));
-      setPresetActions(preset.actions || []);
-    } else {
+    if (!preset) {
       setPresetActions([]);
+      return;
     }
+    setForm((prev) => ({
+      ...prev,
+      name: preset.name,
+      description: preset.description,
+      category: preset.category,
+      theme: preset.theme,
+      status: preset.status || prev.status,
+    }));
+    setPresetActions(preset.actions || []);
   };
 
   const handleChange = (field) => (e) => setForm((prev) => ({ ...prev, [field]: e.target.value }));
@@ -113,78 +94,39 @@ export default function ChallengeForm() {
   const handleSubmit = async (e) => {
     e.preventDefault();
     const payload = { ...form, groupId: form.groupId || null };
+
     if (isEdit) {
       await updateChallenge(Number(id), payload);
       await logActivity(user.id, "Updated challenge", `Updated ${payload.name}`);
-    } else {
-      const newChallenge = await createChallenge({
-        ...payload,
-        createdBy: user.id,
-        actionIds: [],
-        participants: [],
-        joinBy: null,
-      });
-      if (presetActions.length > 0) {
-        const newActionIds = [];
-        for (const tmpl of presetActions) {
-          const newAction = await createAction({ ...tmpl });
-          newActionIds.push(newAction.id);
-        }
-        await updateChallenge(newChallenge.id, { actionIds: newActionIds });
-      }
-      await logActivity(user.id, "Created challenge", `Created ${payload.name}`);
+      navigate("/challenges");
+      return;
     }
+
+    const newChallenge = await createChallenge({
+      ...payload,
+      createdBy: user.id,
+      actionIds: [],
+      participants: [],
+      joinBy: null,
+    });
+
+    // If the user picked a preset, materialize each template action into
+    // a real action row and wire it to the new challenge.
+    if (presetActions.length > 0) {
+      const newActionIds = [];
+      for (const tmpl of presetActions) {
+        const newAction = await createAction({ ...tmpl });
+        newActionIds.push(newAction.id);
+      }
+      await updateChallenge(newChallenge.id, { actionIds: newActionIds });
+    }
+
+    await logActivity(user.id, "Created challenge", `Created ${payload.name}`);
     navigate("/challenges");
   };
 
-  const loadActions = async () => {
+  const reloadActions = async () => {
     if (isEdit) setActions(await getActionsByChallenge(Number(id)));
-  };
-
-  const openNewAction = () => {
-    setEditingAction(null);
-    setActionForm(EMPTY_ACTION);
-    setActionDialogOpen(true);
-  };
-
-  const openEditAction = (action) => {
-    setEditingAction(action);
-    setActionForm({
-      name: action.name,
-      description: action.description,
-      category: action.category,
-      points: action.points,
-    });
-    setActionDialogOpen(true);
-  };
-
-  const handleActionSave = async () => {
-    if (editingAction) {
-      await updateAction(editingAction.id, actionForm);
-    } else {
-      const newAction = await createAction({ ...actionForm });
-      // Add the new action's ID to this challenge's actionIds
-      const challenge = await getChallengeById(Number(id));
-      if (challenge) {
-        await updateChallenge(Number(id), {
-          actionIds: [...(challenge.actionIds || []), newAction.id],
-        });
-      }
-    }
-    setActionDialogOpen(false);
-    loadActions();
-  };
-
-  const handleActionDelete = async (actionId) => {
-    await deleteAction(actionId);
-    // Remove from challenge's actionIds
-    const challenge = await getChallengeById(Number(id));
-    if (challenge) {
-      await updateChallenge(Number(id), {
-        actionIds: (challenge.actionIds || []).filter((aid) => aid !== actionId),
-      });
-    }
-    loadActions();
   };
 
   return (
@@ -204,161 +146,18 @@ export default function ChallengeForm() {
 
       <Paper sx={{ p: { xs: 2, sm: 3 }, maxWidth: 700, mb: 3 }}>
         <form onSubmit={handleSubmit}>
-          <Grid container spacing={2}>
-            {!isEdit && presets.length > 0 && (
-              <Grid size={{ xs: 12 }}>
-                <Paper variant="outlined" sx={{ p: 2, mb: 1, bgcolor: "#f9f9f9" }}>
-                  <Typography variant="subtitle2" color="primary" gutterBottom>
-                    Quick Start: Select a Template
-                  </Typography>
-                  <TextField
-                    select
-                    fullWidth
-                    size="small"
-                    value={selectedPresetId}
-                    onChange={handleApplyPreset}
-                    label="Choose a Preset"
-                  >
-                    <MenuItem value="">None</MenuItem>
-                    {presets.map((p) => (
-                      <MenuItem key={p.id} value={p.id}>
-                        {p.name} ({p.category})
-                      </MenuItem>
-                    ))}
-                  </TextField>
-                </Paper>
-              </Grid>
-            )}
-
-            <Grid size={{ xs: 12 }}>
-              <TextField
-                label="Challenge Name"
-                value={form.name}
-                onChange={handleChange("name")}
-                required
-                fullWidth
+          {!isEdit && (
+            <Box mb={2}>
+              <PresetPicker
+                presets={presets}
+                selectedPresetId={selectedPresetId}
+                onSelect={handleApplyPreset}
+                presetActions={presetActions}
               />
-            </Grid>
-            <Grid size={{ xs: 12 }}>
-              <TextField
-                label="Description"
-                value={form.description}
-                onChange={handleChange("description")}
-                multiline
-                rows={3}
-                fullWidth
-              />
-            </Grid>
-            <Grid size={{ xs: 12, sm: 6 }}>
-              <TextField
-                label="Category"
-                select
-                value={form.category}
-                onChange={handleChange("category")}
-                fullWidth
-              >
-                {CATEGORIES.map((c) => (
-                  <MenuItem key={c} value={c}>
-                    {c}
-                  </MenuItem>
-                ))}
-              </TextField>
-            </Grid>
-            <Grid size={{ xs: 12, sm: 6 }}>
-              <TextField
-                label="Status"
-                select
-                value={form.status}
-                onChange={handleChange("status")}
-                fullWidth
-              >
-                {Object.values(CHALLENGE_STATUSES).map((s) => (
-                  <MenuItem key={s} value={s}>
-                    {s}
-                  </MenuItem>
-                ))}
-              </TextField>
-            </Grid>
-            <Grid size={{ xs: 12, sm: 6 }}>
-              <TextField
-                label="Start Date"
-                type="date"
-                value={form.startDate}
-                onChange={handleChange("startDate")}
-                InputLabelProps={{ shrink: true }}
-                required
-                fullWidth
-              />
-            </Grid>
-            <Grid size={{ xs: 12, sm: 6 }}>
-              <TextField
-                label="End Date"
-                type="date"
-                value={form.endDate}
-                onChange={handleChange("endDate")}
-                InputLabelProps={{ shrink: true }}
-                required
-                fullWidth
-              />
-            </Grid>
-            <Grid size={{ xs: 12, sm: 6 }}>
-              <TextField
-                label="Theme Color"
-                type="color"
-                value={form.theme}
-                onChange={handleChange("theme")}
-                fullWidth
-                InputLabelProps={{ shrink: true }}
-              />
-            </Grid>
-            <Grid size={{ xs: 12, sm: 6 }}>
-              <TextField
-                label="Group"
-                select
-                value={form.groupId || ""}
-                onChange={handleChange("groupId")}
-                fullWidth
-              >
-                <MenuItem value="">No Group</MenuItem>
-                {groups.map((g) => (
-                  <MenuItem key={g.id} value={g.id}>
-                    {g.name}
-                  </MenuItem>
-                ))}
-              </TextField>
-            </Grid>
-          </Grid>
-
-          {!isEdit && presetActions.length > 0 && (
-            <Box mt={3}>
-              <Typography variant="subtitle2" fontWeight={600} mb={1}>
-                Actions from Preset ({presetActions.length})
-              </Typography>
-              <Typography variant="caption" color="text.secondary" display="block" mb={1}>
-                These actions will be created automatically when you submit.
-              </Typography>
-              <TableContainer component={Paper} variant="outlined" sx={{ overflowX: "auto" }}>
-                <Table size="small" sx={{ minWidth: 400 }}>
-                  <TableHead>
-                    <TableRow>
-                      <TableCell>Name</TableCell>
-                      <TableCell>Category</TableCell>
-                      <TableCell align="right">Points</TableCell>
-                    </TableRow>
-                  </TableHead>
-                  <TableBody>
-                    {presetActions.map((a, i) => (
-                      <TableRow key={i}>
-                        <TableCell>{a.name}</TableCell>
-                        <TableCell>{a.category}</TableCell>
-                        <TableCell align="right">{a.points}</TableCell>
-                      </TableRow>
-                    ))}
-                  </TableBody>
-                </Table>
-              </TableContainer>
             </Box>
           )}
+
+          <ChallengeFieldsSection form={form} onChange={handleChange} groups={groups} />
 
           <Stack direction="row" spacing={2} mt={3}>
             <Button type="submit" variant="contained">
@@ -372,112 +171,12 @@ export default function ChallengeForm() {
       </Paper>
 
       {isEdit && (
-        <Box>
-          <Stack direction="row" justifyContent="space-between" alignItems="center" mb={1}>
-            <Typography variant="h6" fontWeight={600}>
-              Actions ({actions.length})
-            </Typography>
-            <Button size="small" startIcon={<AddIcon />} onClick={openNewAction}>
-              Add Action
-            </Button>
-          </Stack>
-          <TableContainer component={Paper} sx={{ overflowX: "auto" }}>
-            <Table size="small" sx={{ minWidth: 400 }}>
-              <TableHead>
-                <TableRow>
-                  <TableCell>Name</TableCell>
-                  <TableCell>Category</TableCell>
-                  <TableCell align="right">Points</TableCell>
-                  <TableCell align="right">Manage</TableCell>
-                </TableRow>
-              </TableHead>
-              <TableBody>
-                {actions.map((a) => (
-                  <TableRow key={a.id}>
-                    <TableCell>{a.name}</TableCell>
-                    <TableCell>{a.category}</TableCell>
-                    <TableCell align="right">{a.points}</TableCell>
-                    <TableCell align="right">
-                      <IconButton size="small" onClick={() => openEditAction(a)}>
-                        <EditIcon fontSize="small" />
-                      </IconButton>
-                      <IconButton
-                        size="small"
-                        color="error"
-                        onClick={() => handleActionDelete(a.id)}
-                      >
-                        <DeleteIcon fontSize="small" />
-                      </IconButton>
-                    </TableCell>
-                  </TableRow>
-                ))}
-                {actions.length === 0 && (
-                  <TableRow>
-                    <TableCell colSpan={4} align="center">
-                      No actions yet. Add one above.
-                    </TableCell>
-                  </TableRow>
-                )}
-              </TableBody>
-            </Table>
-          </TableContainer>
-        </Box>
+        <ActionsEditor
+          challengeId={Number(id)}
+          actions={actions}
+          onChanged={reloadActions}
+        />
       )}
-
-      <Dialog
-        open={actionDialogOpen}
-        onClose={() => setActionDialogOpen(false)}
-        maxWidth="sm"
-        fullWidth
-      >
-        <DialogTitle>{editingAction ? "Edit Action" : "Add Action"}</DialogTitle>
-        <DialogContent>
-          <Stack spacing={2} mt={1}>
-            <TextField
-              label="Action Name"
-              value={actionForm.name}
-              onChange={(e) => setActionForm((p) => ({ ...p, name: e.target.value }))}
-              required
-              fullWidth
-            />
-            <TextField
-              label="Description"
-              value={actionForm.description}
-              onChange={(e) => setActionForm((p) => ({ ...p, description: e.target.value }))}
-              multiline
-              rows={2}
-              fullWidth
-            />
-            <TextField
-              label="Category"
-              select
-              value={actionForm.category}
-              onChange={(e) => setActionForm((p) => ({ ...p, category: e.target.value }))}
-              fullWidth
-            >
-              {CATEGORIES.map((c) => (
-                <MenuItem key={c} value={c}>
-                  {c}
-                </MenuItem>
-              ))}
-            </TextField>
-            <TextField
-              label="Points"
-              type="number"
-              value={actionForm.points}
-              onChange={(e) => setActionForm((p) => ({ ...p, points: Number(e.target.value) }))}
-              fullWidth
-              inputProps={{ min: 1 }}
-            />
-          </Stack>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={() => setActionDialogOpen(false)}>Cancel</Button>
-          <Button variant="contained" onClick={handleActionSave}>
-            {editingAction ? "Save" : "Add"}
-          </Button>
-        </DialogActions>
-      </Dialog>
     </Box>
   );
 }
